Separate normal end-to-end tests and stress tests
